+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 * Skip freeing cursors that were not loaded
 * Fix sound mixing for uncontrolled time
 * Fix concurrent access to action queue causing a crash in libTAS
+* Vulkan functions loaded through dlsym are correctly hooked
 
 ## [1.4.8] - 2026-05-14
 ### Added